+++ frontend/src/mocks/handlers.ts
@@ -14,38 +14,11 @@ import runGroupsFixture from '../../fixtures/run_groups.json';
 import runRequestDetailsFixture from '../../fixtures/run_request_details.json';
 import runStatusRunnerFixture from '../../fixtures/run_status_runner.json';
 import runStatusShellFixture from '../../fixtures/run_status_shell.json';
-import runStatusShellPlaylistFixture from '../../fixtures/run_status_shell_playlist.json';
-
-// Mutable copy of the playlist shell so playlist edits/retries/finalises made in mock mode stick
-// across refetches. The templates keep the original fixture's queued-run shape for added tests.
-const playlistShell = structuredClone(runStatusShellPlaylistFixture);
-const upcomingRunTemplate = runStatusShellPlaylistFixture.playlist_runs[2];
-const upcomingSummaryTemplate = runStatusShellPlaylistFixture.run.playlist_runs[2];
-
-const syncPlaylistSummary = () => {
-  playlistShell.run.playlist_runs = playlistShell.playlist_runs.map(
-    ({ run_id, test_procedure_id, status }) => ({ run_id, test_procedure_id, status })
-  );
-};
-
-// Compose the shell for one run of the mock playlist, so navigating between playlist runs works.
-const playlistShellForRun = (runId: number) => {
-  const index = playlistShell.playlist_runs.findIndex((r) => r.run_id === runId);
-  if (index === -1) {
-    return playlistShell;
-  }
-  const row = playlistShell.playlist_runs[index];
-  return {
-    ...playlistShell,
-    run: {
-      ...playlistShell.run,
-      ...row,
-      playlist_order: index,
-      playlist_runs: playlistShell.run.playlist_runs,
-    },
-    run_is_live: row.status === 'started' || row.status === 'provisioning',
-  };
-};
+import {
+  finaliseMockPlaylistRun,
+  playlistShellForRun,
+  replaceMockPlaylistTail,
+} from './playlistShell';
 import sessionFixture from '../../fixtures/session.json';
 import sessionAdminFixture from '../../fixtures/session_admin.json';
 import complianceRequestsFixture from '../../fixtures/compliance_requests.json';
@@ -125,18 +98,7 @@ export const handlers = [
     HttpResponse.json({ run_id: Number(params.runId) })
   ),
   http.post('/api/runs/:runId/finalise', ({ params }) => {
-    // Finalising a mock playlist run records it and hands over to the next queued test.
-    const row = playlistShell.playlist_runs.find((r) => r.run_id === Number(params.runId));
-    if (row) {
-      row.status = 'finalised';
-      row.all_criteria_met = row.all_criteria_met ?? false;
-      row.finalised_at = new Date().toISOString();
-      const next = playlistShell.playlist_runs.find((r) => r.status === 'initialised');
-      if (next) {
-        next.status = 'started';
-      }
-      syncPlaylistSummary();
-    }
+    finaliseMockPlaylistRun(Number(params.runId));
     return HttpResponse.json({ run_id: Number(params.runId) });
   }),
   http.delete('/api/runs/:runId', ({ params }) =>
@@ -164,23 +126,7 @@ export const handlers = [
   ),
   http.post('/api/run/:runId/playlist', async ({ request }) => {
     const body = (await request.json()) as { test_procedure_ids: string[] };
-    const summaryTail = body.test_procedure_ids.map((test_procedure_id, i) => ({
-      ...upcomingSummaryTemplate,
-      run_id: 900 + i,
-      test_procedure_id,
-    }));
-    // Reflect the new tail in the mock shell so the banner shows the edit on refetch.
-    playlistShell.playlist_runs = [
-      ...playlistShell.playlist_runs.filter((r) => r.status !== 'initialised'),
-      ...summaryTail.map((s) => ({
-        ...upcomingRunTemplate,
-        run_id: s.run_id,
-        test_procedure_id: s.test_procedure_id,
-        test_url: `https://cactus.example/run/${s.run_id}`,
-      })),
-    ];
-    syncPlaylistSummary();
-    return HttpResponse.json({ playlist_runs: summaryTail });
+    return HttpResponse.json({ playlist_runs: replaceMockPlaylistTail(body.test_procedure_ids) });
   }),
   http.post('/api/runs/:runId/proceed', () => HttpResponse.json({ handled: true })),
   http.post('/api/admin/runs/:runId/proceed', () => HttpResponse.json({ handled: true })),
